chore: refactored api and controller
