+++ src/main/java/de/fraunhofer/isst/configmanager/controller/ResourceRepresentationApi.java
@@ -34,8 +34,7 @@ public interface ResourceRepresentationApi {
     @GetMapping(value = "/resource/representation", produces = "application/ld+json")
     @Operation(summary = "Get the representation for a resource")
     @ApiResponses({@ApiResponse(responseCode = "200", description = "Successfully get the representation for a resource")})
-    ResponseEntity<String> getResourceRepresentation(@RequestParam("resourceId") URI resourceId,
-                                                     @RequestParam("representationId") URI representationId);
+    ResponseEntity<String> getResourceRepresentation(@RequestParam("representationId") URI representationId);
 
     @GetMapping(value = "/resource/representation/json", produces = "application/ld+json")
     @Operation(summary = "Get the representation for a resource in JSON format")
