+++ src/main/java/de/fraunhofer/isst/configmanager/controller/ResourceRepresentationUIController.java
@@ -7,6 +7,7 @@ import de.fraunhofer.iais.eis.util.Util;
 import de.fraunhofer.isst.configmanager.communication.clients.DefaultConnectorClient;
 import de.fraunhofer.isst.configmanager.configmanagement.service.ConfigModelService;
 import de.fraunhofer.isst.configmanager.configmanagement.service.RepresentationEndpointService;
+import de.fraunhofer.isst.configmanager.configmanagement.service.ResourceService;
 import de.fraunhofer.isst.configmanager.configmanagement.service.UtilService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import net.minidev.json.JSONObject;
@@ -21,7 +22,6 @@ import org.springframework.web.bind.annotation.RestController;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.net.URI;
-import java.util.Collection;
 
 /**
  * The controller class implements the ResourceRepresentationApi and offers the possibilities to manage
@@ -36,19 +36,22 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
 
     private final ConfigModelService configModelService;
     private final UtilService utilService;
+    private final RepresentationEndpointService representationEndpointService;
+    private final ResourceService resourceService;
     private final DefaultConnectorClient client;
     private final Serializer serializer;
-    private final RepresentationEndpointService representationEndpointService;
 
     @Autowired
     public ResourceRepresentationUIController(ConfigModelService configModelService,
                                               UtilService utilService,
+                                              ResourceService resourceService,
                                               DefaultConnectorClient client,
                                               Serializer serializer,
                                               RepresentationEndpointService representationEndpointService) {
         this.client = client;
         this.configModelService = configModelService;
         this.utilService = utilService;
+        this.resourceService = resourceService;
         this.serializer = serializer;
         this.representationEndpointService = representationEndpointService;
     }
@@ -100,15 +103,12 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
 
             var response = client.registerResourceRepresentation(resourceId.toString(), representation);
             jsonObject.put("connectorResponse", response);
-
             // Updates the custom resource representation of the connector
             ResponseEntity<String> res =
                     utilService.addEndpointToConnectorRepresentation(endpointId, resourceId, representation);
             logger.info("Response of updates custom resource representation: {}", res);
-
             // Saves the endpoint id associated with the representation id in the database
             representationEndpointService.createRepresentationEndpoint(endpointId, representation.getId());
-
             return ResponseEntity.ok(jsonObject.toJSONString());
         } catch (IOException e) {
             logger.error(e.getMessage());
@@ -141,46 +141,21 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
         }
 
         // Determine the correct resource representation from the resource catalog of the connector
-        RepresentationImpl catalogCandidate = (RepresentationImpl) configModelService.getConfigModel()
-                .getConnectorDescription().getResourceCatalog().stream()
-                .map(ResourceCatalog::getOfferedResource)
-                .flatMap(Collection::stream)
-                .map(DigitalContent::getRepresentation)
-                .flatMap(Collection::stream)
-                .filter(representation -> representation.getId().equals(representationId))
-                .findAny()
-                .orElse(null);
-
-        // Check if parameters are null and when not update it with new values
-        updateRepresentation(language, filenameExtension, bytesize, sourceType, catalogCandidate);
+        RepresentationImpl catalogCandidate = resourceService.getResourceRepresentationInCatalog(representationId);
+        if (catalogCandidate != null) {
+            // Check if parameters are null and when not update it with new values
+            resourceService.updateRepresentation(language, filenameExtension, bytesize, sourceType, catalogCandidate);
+        }
 
         // Try to update resource representation in the app routes, if it exists
         if (configModelService.getConfigModel().getAppRoute() == null) {
             logger.info("Could not find any app route to update the resource representation");
         } else {
-            RepresentationImpl appRouteCandidate = null;
-            for (AppRoute appRoute : configModelService.getConfigModel().getAppRoute()) {
-                if (appRoute.getHasSubRoute() != null) {
-                    for (RouteStep routeStep : appRoute.getHasSubRoute()) {
-                        if (routeStep.getAppRouteOutput() != null) {
-                            for (Resource resource : routeStep.getAppRouteOutput()) {
-                                if (resourceId.equals(resource.getId())) {
-                                    if (resource.getRepresentation() != null) {
-                                        for (Representation representation : resource.getRepresentation()) {
-                                            if (representationId.equals(representation.getId())) {
-                                                appRouteCandidate = (RepresentationImpl) representation;
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
+            RepresentationImpl appRouteCandidate = resourceService.getResourceRepresentationInAppRoute(resourceId, representationId);
             // Check if parameters are null and when not update it with new values
-            updateRepresentation(language, filenameExtension, bytesize, sourceType, appRouteCandidate);
+            if (appRouteCandidate != null) {
+                resourceService.updateRepresentation(language, filenameExtension, bytesize, sourceType, appRouteCandidate);
+            }
         }
 
         try {
@@ -216,34 +191,29 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
     /**
      * This method returns the specific representation from a resource with the given parameters.
      *
-     * @param resourceId       id of the resource
      * @param representationId id of the representation
      * @return a suitable http response depending on success
      */
     @Override
-    public ResponseEntity<String> getResourceRepresentation(URI resourceId, URI representationId) {
+    public ResponseEntity<String> getResourceRepresentation(URI representationId) {
 
         if (configModelService.getConfigModel() == null ||
                 configModelService.getConfigModel().getConnectorDescription().getResourceCatalog() == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Could not find any resources!\"}");
         }
-        for (ResourceCatalog resourceCatalog : configModelService.getConfigModel().getConnectorDescription().getResourceCatalog()) {
-            for (Resource resource : resourceCatalog.getOfferedResource()) {
-                if (resourceId.equals(resource.getId())) {
-                    for (Representation representation : resource.getRepresentation()) {
-                        if (representationId.equals(representation.getId())) {
-                            try {
-                                return ResponseEntity.ok(serializer.serialize(representation));
-                            } catch (IOException e) {
-                                logger.error(e.getMessage());
-                            }
-
-                        }
-                    }
-                }
+        RepresentationImpl representation = resourceService.getResourceRepresentationInCatalog(representationId);
+        if (representation != null) {
+            try {
+                return ResponseEntity.ok(serializer.serialize(representation));
+            } catch (IOException e) {
+                logger.error(e.getMessage());
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Problems while serializing the " +
+                        "representation");
             }
+        } else {
+            return ResponseEntity.badRequest().body("Could not get resource representation");
+
         }
-        return ResponseEntity.badRequest().body("Could not get resource representation");
     }
 
     /**
@@ -294,47 +264,9 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Could not find any resources!\"}");
         }
 
-        var deleted = false;
-
-        // Delete representation in catalog
-        for (ResourceCatalog resourceCatalog : configModelService.getConfigModel().getConnectorDescription().getResourceCatalog()) {
-            if (resourceCatalog != null) {
-                var resource = resourceCatalog.getOfferedResource().stream()
-                        .filter(resource1 -> resource1.getId().equals(resourceId)).findAny().orElse(null);
-                if (resource != null) {
-                    deleted |= resource.getRepresentation()
-                            .removeIf(representation -> representation.getId().equals(representationId));
-                }
-            }
-        }
-
-        // Delete representation in app route if exists
-        if (configModelService.getConfigModel().getAppRoute() == null) {
-            logger.info("No app route found to delete the resource representation");
-        } else {
-            Resource foundresource = null;
-            for (AppRoute appRoute : configModelService.getConfigModel().getAppRoute()) {
-                if (appRoute.getHasSubRoute() != null) {
-                    for (RouteStep routeStep : appRoute.getHasSubRoute()) {
-                        if (routeStep.getAppRouteOutput() != null) {
-                            for (Resource resource : routeStep.getAppRouteOutput()) {
-                                if (resourceId.equals(resource.getId())) {
-                                    foundresource = resource;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            if (foundresource != null) {
-                deleted |= foundresource.getRepresentation()
-                        .removeIf(representation -> representation.getId().equals(representationId));
-            }
-
-        }
-        try {
-            if (deleted) {
+        boolean deleted = resourceService.deleteResourceRepresentation(resourceId, representationId);
+        if (deleted) {
+            try {
                 var response = client.deleteResourceRepresentation(resourceId.toString(),
                         representationId.toString());
                 configModelService.saveState();
@@ -343,41 +275,11 @@ public class ResourceRepresentationUIController implements ResourceRepresentatio
                 jsonObject.put("resourceID", resourceId.toString());
                 jsonObject.put("representationID", representationId.toString());
                 return ResponseEntity.ok(jsonObject.toJSONString());
-            } else {
-                return ResponseEntity.notFound().build();
+            } catch (IOException e) {
+                return ResponseEntity.badRequest().body("Problems while deleting the representation at the connector");
             }
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-        return ResponseEntity.badRequest().body("Could not delete the resource representation");
-    }
-
-    /**
-     * @param language           language of the representation
-     * @param filenameExtension  filename extension
-     * @param bytesize           byte size of the representation
-     * @param sourceType         source type of the representation
-     * @param representationImpl representation implementation
-     */
-    private void updateRepresentation(String language, String filenameExtension, Long bytesize, String sourceType,
-                                      RepresentationImpl representationImpl) {
-        if (representationImpl != null) {
-            if (language != null) {
-                representationImpl.setLanguage(Language.valueOf(language));
-            }
-            if (filenameExtension != null) {
-                representationImpl.setMediaType(null);
-                representationImpl.setMediaType(new IANAMediaTypeBuilder()
-                        ._filenameExtension_(filenameExtension).build());
-            }
-            if (bytesize != null) {
-                representationImpl.setInstance(null);
-                representationImpl.setInstance(Util.asList(new ArtifactBuilder()
-                        ._byteSize_(BigInteger.valueOf(bytesize)).build()));
-            }
-            if (sourceType != null) {
-                representationImpl.setProperty("ids:sourceType", sourceType);
-            }
+        } else {
+            return ResponseEntity.badRequest().body("Could not delete the resource representation");
         }
     }
 }
